# clients/java/api/openapi.yaml
openapi: 3.1.0
info:
  description: "Define, evaluate, and enforce data retention policies"
  title: DRPE — Data Retention Policy Engine
  version: 0.1.0
servers:
- description: Local DRPE API
  url: http://localhost:8000
paths:
...
  /api/v1/webhooks:
    get:
      operationId: list_webhooks_api_v1_webhooks_get
      parameters:
      - explode: true
        in: query
        name: active
        required: false
        schema:
          nullable: true
          type: boolean
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
        style: form
      - explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/WebhookResponse"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: List Webhooks
      tags:
      - webhooks
      x-accepts:
      - application/json
    post:
      operationId: create_webhook_api_v1_webhooks_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookCreateResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Create Webhook
      tags:
      - webhooks
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/webhooks/{webhook_id}:
    delete:
      operationId: delete_webhook_api_v1_webhooks__webhook_id__delete
      parameters:
      - explode: false
        in: path
        name: webhook_id
        required: true
        schema:
          title: Webhook Id
          type: string
        style: simple
      responses:
        "204":
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Delete Webhook
      tags:
      - webhooks
      x-accepts:
      - application/json
    get:
      operationId: get_webhook_api_v1_webhooks__webhook_id__get
      parameters:
      - explode: false
        in: path
        name: webhook_id
        required: true
        schema:
          title: Webhook Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Webhook
      tags:
      - webhooks
      x-accepts:
      - application/json
    patch:
      operationId: update_webhook_api_v1_webhooks__webhook_id__patch
      parameters:
      - explode: false
        in: path
        name: webhook_id
        required: true
        schema:
          title: Webhook Id
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookUpdateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Update Webhook
      tags:
      - webhooks
      x-content-type: application/json
      x-accepts:
      - application/json
...
  /api/v1/privacy/status:
    get:
      operationId: privacy_status_api_v1_privacy_status_get
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyStatusResponse"
          description: Successful Response
      security:
      - BearerAuth: []
      summary: Privacy Status
      tags:
      - privacy
      x-accepts:
      - application/json
  /api/v1/privacy/mask:
    post:
      operationId: privacy_mask_api_v1_privacy_mask_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaskRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaskResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Privacy Mask
      tags:
      - privacy
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/privacy/unmask:
    post:
      operationId: privacy_unmask_api_v1_privacy_unmask_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnmaskRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnmaskResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Privacy Unmask
      tags:
      - privacy
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    MaskRequest:
      example:
        messages:
        - key: messages
        - key: messages
        text: text
      properties:
        text:
          nullable: true
          type: string
        messages:
          items:
            additionalProperties:
              type: string
          nullable: true
          type: array
      title: MaskRequest
    MaskResponse:
      example:
        entity_count: 0
        masked_messages:
        - key: masked_messages
        - key: masked_messages
        mapping_token: mapping_token
        masked_text: masked_text
      properties:
        masked_text:
          nullable: true
          type: string
        masked_messages:
          items:
            additionalProperties:
              type: string
          nullable: true
          type: array
        mapping_token:
          title: Mapping Token
          type: string
        entity_count:
          title: Entity Count
          type: integer
      required:
      - entity_count
      - mapping_token
      title: MaskResponse
...
    PrivacyStatusResponse:
      example:
        languages:
        - languages
        - languages
        available: true
        enabled: true
        model_size: model_size
      properties:
        available:
          title: Available
          type: boolean
        enabled:
          title: Enabled
          type: boolean
        model_size:
          nullable: true
          type: string
        languages:
          items:
            type: string
          type: array
      required:
      - available
      - enabled
      title: PrivacyStatusResponse
...
    UnmaskRequest:
      example:
        consume_token: true
        mapping_token: mapping_token
        text: text
      properties:
        text:
          title: Text
          type: string
        mapping_token:
          title: Mapping Token
          type: string
        consume_token:
          default: true
          title: Consume Token
          type: boolean
      required:
      - mapping_token
      - text
      title: UnmaskRequest
    UnmaskResponse:
      example:
        text: text
      properties:
        text:
          title: Text
          type: string
      required:
      - text
      title: UnmaskResponse
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        input: ""
        ctx: "{}"
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
        input: {}
        ctx:
          title: Context
          type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WebhookCreateRequest:
      description: POST /api/v1/webhooks body.
      example:
        description: description
        active: true
        secret: secret
        url: https://openapi-generator.tech
        events:
        - events
        - events
      properties:
        url:
          format: uri
          minLength: 1
          title: Url
          type: string
        events:
          items:
            type: string
          minItems: 1
          type: array
        secret:
          maxLength: 256
          minLength: 8
          nullable: true
          type: string
        description:
          maxLength: 512
          nullable: true
          type: string
        active:
          default: true
          title: Active
          type: boolean
      required:
      - events
      - url
      title: WebhookCreateRequest
    WebhookCreateResponse:
      description: Create response includes secret once for the client to store.
      example:
        secret_set: false
        updated_at: 2000-01-23T04:56:07.000+00:00
        description: description
        active: true
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        secret: secret
        url: url
        events:
        - events
        - events
      properties:
        id:
          title: Id
          type: string
        url:
          title: Url
          type: string
        events:
          items:
            type: string
          type: array
        description:
          nullable: true
          type: string
        active:
          default: true
          title: Active
          type: boolean
        secret_set:
          default: false
          title: Secret Set
          type: boolean
        created_at:
          format: date-time
          title: Created At
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        secret:
          title: Secret
          type: string
      required:
      - created_at
      - events
      - id
      - secret
      - updated_at
      - url
      title: WebhookCreateResponse
    WebhookResponse:
      description: Public webhook view (secret omitted).
      example:
        secret_set: false
        updated_at: 2000-01-23T04:56:07.000+00:00
        description: description
        active: true
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        url: url
        events:
        - events
        - events
      properties:
        id:
          title: Id
          type: string
        url:
          title: Url
          type: string
        events:
          items:
            type: string
          type: array
        description:
          nullable: true
          type: string
        active:
          default: true
          title: Active
          type: boolean
        secret_set:
          default: false
          title: Secret Set
          type: boolean
        created_at:
          format: date-time
          title: Created At
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - created_at
      - events
      - id
      - updated_at
      - url
      title: WebhookResponse
    WebhookUpdateRequest:
      description: "PATCH /api/v1/webhooks/{id} body."
      example:
        description: description
        active: true
        secret: secret
        url: https://openapi-generator.tech
        events:
        - events
        - events
      properties:
        url:
          format: uri
          minLength: 1
          nullable: true
          type: string
        events:
          items:
            type: string
          minItems: 1
          nullable: true
          type: array
        secret:
          maxLength: 256
          minLength: 8
          nullable: true
          type: string
        description:
          maxLength: 512
          nullable: true
          type: string
        active:
          nullable: true
          type: boolean
      title: WebhookUpdateRequest
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes:
    BearerAuth:
      description: API key as Bearer token (DRPE_API_KEY)
      scheme: bearer
      type: http